server/requirements.txt
pymongo
fastapi
pydantic
virtualenv
uvicorn
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
mongoengine
pyotp
python-dotenv
geopy
passlib
uuid